otlp-experiment/pyproject.toml
[project]
name = "otlp-experiment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "fastapi[standard]>=0.117.1",
    "opentelemetry-api>=1.37.0",
    "opentelemetry-distro>=0.58b0",
    "opentelemetry-exporter-otlp>=1.37.0",
    "opentelemetry-instrumentation>=0.58b0",
    "opentelemetry-instrumentation-asgi>=0.58b0",
    "opentelemetry-instrumentation-fastapi>=0.58b0",
    "opentelemetry-instrumentation-logging>=0.58b0",
    "opentelemetry-sdk>=1.37.0",
]
